Move FeatureManager tests to test_utils
