+++ tests/test_importers/test_utils.py
@@ -1,7 +1,5 @@
-from dataclasses import dataclass
-from unittest.mock import MagicMock, patch
-
 import pytest
+from dataclasses import dataclass
 
 from PyQt5.QtCore import QVariant
 
@@ -171,3 +169,42 @@ def test_get_substring_geometry_argument_processing(simplify):
         assert len(result.asPolyline()) >= 2
 
     assert result.length() == end_distance - start_distance
+
+
+@pytest.fixture
+def node_fields():
+    fields = QgsFields()
+    fields.append(QgsField("id", QVariant.Int))
+    fields.append(QgsField("foo", QVariant.String))
+    return fields
+
+
+@pytest.fixture
+def node_point():
+    return QgsPointXY(1.0, 2.0)
+
+
+@pytest.fixture
+def node_geom(node_point):
+    return QgsGeometry.fromPointXY(node_point)
+
+
+@pytest.mark.parametrize('next_id', [1, 100])
+def test_feature_manager_increment_id(next_id, node_geom, node_fields):
+    manager = FeatureManager(next_id)
+    assert manager.next_id == next_id
+    node_feat = manager.create_new(node_geom, node_fields)
+    assert node_feat["id"] == next_id
+    assert manager.next_id == next_id + 1
+
+
+def test_feature_manager_create_new(node_geom, node_fields):
+    manager = FeatureManager()
+    node_feat = manager.create_new(node_geom, node_fields)
+    assert node_feat.geometry().asWkt() == node_geom.asWkt()
+
+
+def test_feature_manager_create_new_with_attributes(node_geom, node_fields):
+    manager = FeatureManager()
+    node_feat = manager.create_new(node_geom, node_fields, attributes={"foo": "bar"})
+    assert node_feat["foo"] == "bar"
